Fix format string issues flagged in review: 0x%p and UINT_PTR cast
Agent-Logs-Url: https://github.com/dotnet/runtime/sessions/cfaa2eac-6dfc-4654-be66-b9da86ed1c39

+++ src/coreclr/debug/di/rsthread.cpp
@@ -689,7 +689,7 @@ HRESULT CordbThread::SetDebugState(CorDebugThreadState state)
     FAIL_IF_NEUTERED(this);
     ATT_REQUIRE_STOPPED_MAY_FAIL(GetProcess());
 
-    LOG((LF_CORDB, LL_INFO1000, "CT::SDS: thread=%p 0x%zx, state=%d\n", this, m_id, state));
+    LOG((LF_CORDB, LL_INFO1000, "CT::SDS: thread=%p 0x%zx, state=%d\n", this, (size_t)m_id, state));
 
     // @dbgtodo- , sync - decide on how to suspend a thread. V2 leverages synchronization
     // (see below). For V3, do we just hard suspend the thread?
